Add Order service
Add addPaymentDataToOrderFromResponse() and outsource the additional
payment data collection and saving to the order functionality form
FrontController.php handlePaymentsResponse()
